realtime-agent/requirements.txt
fastapi==0.127.0
uvicorn==0.40.0
python-dotenv==1.2.1
httpx==0.28.1
